Fixed CI build

+++ src/fetch-plugins.ts
@@ -11,12 +11,15 @@ export module FetchPlugins {
       let src = `node_modules/${plugin}`;
       if (!fs.pathExistsSync(src)) {
         console.log(`Installing plugin via npm: ${fetch[plugin].source.id}`);
-        if (shell.exec(`npm install ${fetch[plugin].source.id} --no-save`).code !== 0) {
+        let execOutput = shell.exec(`npm install ${fetch[plugin].source.id} --no-save`);
+        if (execOutput !== null && execOutput.code !== 0) {
           console.log(`Skipping plugin ${plugin}`);
           continue;
         }
       }
-      fs.copySync(src, `plugins/${plugin}`);
+      if (fs.pathExistsSync(src)) {
+        fs.copySync(src, `plugins/${plugin}`);
+      }
     }
   }
